requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
streamlit
requests
streamlit-code-editor
google-generativeai
